move async logic to the walletconnect store
