+++ src/components/walletconnect/WC2TransactionRequest.vue
@@ -1,21 +1,12 @@
 <script setup lang="ts">
-  import { ref, toRefs } from 'vue';
+  import { toRefs } from 'vue';
+  import { binToHex, lockingBytecodeToCashAddress, type TransactionCommon, type Input, type Output } from "@bitauth/libauth"
   import { useDialogPluginComponent } from 'quasar'
-  import { lockingBytecodeToCashAddress, hexToBin, binToHex, importWalletTemplate, walletTemplateP2pkhNonHd, walletTemplateToCompilerBCH, secp256k1, generateTransaction, encodeTransaction, sha256, hash256, SigningSerializationFlag, generateSigningSerializationBCH, TransactionCommon, TransactionTemplateFixed, CompilationContextBCH, Input, Output } from "@bitauth/libauth"
   import { BCMR, convert } from "mainnet-js"
-  import { getSdkError } from '@walletconnect/utils';
-  import alertDialog from 'src/components/alertDialog.vue'
   import type { DappMetadata, ContractInfo } from "src/interfaces/interfaces"
   import { useStore } from 'src/stores/store'
-  import { useWalletconnectStore } from 'src/stores/walletconnectStore'
   import { parseExtendedJson } from 'src/utils/utils'
-  import { useQuasar } from 'quasar'
-  const $q = useQuasar()
   const store = useStore()
-  const walletconnectStore = useWalletconnectStore()
-  const web3wallet = walletconnectStore.web3wallet
-
-  const showDialog = ref(true);
 
   const props = defineProps<{
     dappMetadata: DappMetadata,
@@ -27,11 +18,7 @@
     ...useDialogPluginComponent.emits
   ])
   const { dialogRef, onDialogHide, onDialogOK, onDialogCancel } = useDialogPluginComponent()
-  function onOKClick () {
-    onDialogOK()
-  }
-
-  const { id, topic } = transactionRequestWC.value;
+  
   // parse params from transactionRequestWC
   const requestParams = parseExtendedJson(JSON.stringify(transactionRequestWC.value.params.request.params));
   const txDetails:TransactionCommon = requestParams.transaction;
@@ -90,129 +77,10 @@
       else tokensReceivedOutputs[tokenCategory] = [output.token];
     }
   }
-
-
-  async function signTransactionWC() {
-    const privateKey = store?.wallet?.privateKey;
-    const pubkeyCompressed = store?.wallet?.publicKeyCompressed
-    if(!privateKey || !pubkeyCompressed) return
-
-    // prepare libauth template for input signing
-    const template = importWalletTemplate(walletTemplateP2pkhNonHd);
-    if (typeof template === "string") throw new Error("Transaction template error");
-
-    // configure compiler
-    const compiler = walletTemplateToCompilerBCH(template);
-
-    const txTemplate = {...txDetails} as TransactionTemplateFixed<typeof compiler>;
-
-    for (const [index, input] of txTemplate.inputs.entries()) {
-      const sourceOutputsUnpacked = sourceOutputs;
-      if (sourceOutputsUnpacked[index].contract?.artifact.contractName) {
-        // instruct compiler to produce signatures for relevant contract inputs
-
-        // replace pubkey and sig placeholders
-        let unlockingBytecodeHex = binToHex(sourceOutputsUnpacked[index].unlockingBytecode);
-        const sigPlaceholder = "41" + binToHex(Uint8Array.from(Array(65)));
-        const pubkeyPlaceholder = "21" + binToHex(Uint8Array.from(Array(33)));
-        if (unlockingBytecodeHex.indexOf(sigPlaceholder) !== -1) {
-          // compute the signature argument
-          const hashType = SigningSerializationFlag.allOutputs | SigningSerializationFlag.utxos | SigningSerializationFlag.forkId;
-          const context: CompilationContextBCH = { inputIndex: index, sourceOutputs: sourceOutputsUnpacked, transaction: txDetails };
-          const signingSerializationType = new Uint8Array([hashType]);
-
-          const coveredBytecode = sourceOutputsUnpacked[index].contract?.redeemScript;
-          if (!coveredBytecode) {
-            alert("Not enough information provided, please include contract redeemScript");
-            return;
-          }
-          const sighashPreimage = generateSigningSerializationBCH(context, { coveredBytecode, signingSerializationType });
-          const sighash = hash256(sighashPreimage);
-          const signature = secp256k1.signMessageHashSchnorr(privateKey, sighash);
-          if (typeof signature === "string") {
-            alert(signature);
-            return;
-          }
-          const sig = Uint8Array.from([...signature, hashType]);
-
-          unlockingBytecodeHex = unlockingBytecodeHex.replace(sigPlaceholder, "41" + binToHex(sig));
-        }
-        if (unlockingBytecodeHex.indexOf(pubkeyPlaceholder) !== -1) {
-          unlockingBytecodeHex = unlockingBytecodeHex.replace(pubkeyPlaceholder, "21" + binToHex(pubkeyCompressed));
-        }
-
-        input.unlockingBytecode = hexToBin(unlockingBytecodeHex);
-      } else {
-        // replace unlocking bytecode for non-contract inputs having placeholder unlocking bytecode
-        const sourceOutput = sourceOutputsUnpacked[index];
-        if (!sourceOutput.unlockingBytecode?.length && toCashaddr(sourceOutput.lockingBytecode) === store.wallet?.getDepositAddress()) {
-           input.unlockingBytecode = {
-            compiler,
-             data: {
-              keys: { privateKeys: { key: privateKey } },
-            },
-            valueSatoshis: sourceOutput.valueSatoshis,
-            script: "unlock",
-            token: sourceOutput.token,
-          }
-        }
-      }
-    };
-
-    // generate and encode transaction
-    const generated = generateTransaction(txTemplate);
-    if (!generated.success) throw Error(JSON.stringify(generated.errors, null, 2));
-
-    const encoded = encodeTransaction(generated.transaction);
-    const hash = binToHex(sha256.hash(sha256.hash(encoded)).reverse());
-    const signedTxObject = { signedTransaction: binToHex(encoded), signedTransactionHash: hash };
-
-    // send transaction
-    const response = { id, jsonrpc: '2.0', result: signedTxObject };
-    if (requestParams.broadcast) {
-      try{
-        $q.notify({
-          spinner: true,
-          message: 'Sending transaction...',
-          color: 'grey-5',
-          timeout: 750
-        })
-        const txId = await store.wallet?.submitTransaction(hexToBin(signedTxObject.signedTransaction));
-        showDialog.value = false
-        const alertMessage = `Sent WalletConnect transaction '${requestParams.userPrompt}'`
-        $q.dialog({
-        component: alertDialog,
-        componentProps: {
-          alertInfo: { message: alertMessage, txid: txId as string } 
-        }
-      })
-      } catch(error){
-        showDialog.value = false
-        console.log(error)
-        $q.notify({
-          type: 'negative',
-          message: typeof error == 'string' ? error :  "Error in sending transaction"
-        })
-        return
-      }   
-    }
-    await web3wallet?.respondSessionRequest({ topic, response });
-    // close dialog when broadcast = false
-    showDialog.value = false
-    // TODO: rework
-    // emit('signedTransaction', requestParams.broadcast);
-  }
-
-  async function rejectTransaction(){
-    const response = { id, jsonrpc: '2.0', error: getSdkError('USER_REJECTED') };
-    await web3wallet?.respondSessionRequest({ topic, response });
-    // TODO: rework
-    // emit('rejectTransaction')
-  }
 </script>
 
 <template>
-  <q-dialog v-model="showDialog" persistent transition-show="scale" transition-hide="scale">
+  <q-dialog ref="dialogRef" @hide="onDialogHide" persistent transition-show="scale" transition-hide="scale">
     <q-card>
       <fieldset class="dialogFieldsetTxRequest"> 
         <legend style="font-size: large;">Sign Transaction</legend>
@@ -334,8 +202,8 @@
           </div>
         </details>
         <div class="wc-modal-bottom-buttons">
-        <input type="button" class="primaryButton" value="Sign" @click="() => signTransactionWC()">
-          <input type="button" value="Cancel" @click="() => rejectTransaction()" v-close-popup>
+          <input type="button" class="primaryButton" value="Sign" @click="onDialogOK">
+          <input type="button" value="Cancel" @click="onDialogCancel" v-close-popup>
         </div>
       </fieldset>
     </q-card>
