+++ src/stores/walletconnectStore.ts
@@ -4,9 +4,35 @@ import { Core } from '@walletconnect/core'
 import { Web3Wallet, type Web3WalletTypes } from '@walletconnect/web3wallet'
 import type Client from '@walletconnect/web3wallet'
 import type {SessionTypes} from '@walletconnect/types'
+import {
+  hexToBin,
+  binToHex,
+  lockingBytecodeToCashAddress,
+  importWalletTemplate,
+  walletTemplateP2pkhNonHd,
+  walletTemplateToCompilerBCH,
+  secp256k1, generateTransaction,
+  encodeTransaction,
+  sha256,
+  hash256,
+  SigningSerializationFlag,
+  generateSigningSerializationBCH,
+  TransactionTemplateFixed,
+  type CompilationContextBCH,
+  type TransactionCommon,
+  type Input,
+  type Output
+} from "@bitauth/libauth"
+import { getSdkError } from '@walletconnect/utils';
+import { parseExtendedJson } from 'src/utils/utils'
+import alertDialog from 'src/components/alertDialog.vue'
 import { Dialog, Notify } from "quasar";
 import WC2TransactionRequest from 'src/components/walletconnect/WC2TransactionRequest.vue';
 import WC2SignMessageRequest from 'src/components/walletconnect/WCSignMessageRequest.vue'
+import AsyncDialogWrapper from 'src/components/walletconnect/asyncDialogWrapper.vue'
+import { useStore } from 'src/stores/store'
+import { ContractInfo } from "src/interfaces/interfaces"
+const store = useStore()
 
 export const useWalletconnectStore = defineStore('walletconnectStore', () => {
 
@@ -55,8 +81,9 @@ export const useWalletconnectStore = defineStore('walletconnectStore', () => {
         const dappMetadata = session.peer.metadata;
         return await new Promise<void>((resolve, reject) => {
           Dialog.create({
-            component: WC2SignMessageRequest,
+            component: AsyncDialogWrapper,
             componentProps: {
+              component: WC2SignMessageRequest,
               dappMetadata,
               signMessageRequestWC: event
             },
@@ -83,8 +110,9 @@ export const useWalletconnectStore = defineStore('walletconnectStore', () => {
         const dappMetadata = session.peer.metadata;
         return await new Promise<void>((resolve, reject) => {
           Dialog.create({
-            component: WC2TransactionRequest,
+            component: AsyncDialogWrapper,
             componentProps: {
+              component: WC2TransactionRequest,
               dappMetadata,
               transactionRequestWC: event
             },
@@ -122,6 +150,138 @@ export const useWalletconnectStore = defineStore('walletconnectStore', () => {
     })
   }
 */
+
+  const toCashaddr = (lockingBytecode:Uint8Array) => {
+    const prefix = store.network == "mainnet" ? "bitcoincash" : "bchtest";
+    // check for opreturn
+    if(binToHex(lockingBytecode).startsWith("6a")) return "opreturn:" +  binToHex(lockingBytecode)
+    const result = lockingBytecodeToCashAddress(lockingBytecode,prefix);
+    if (typeof result !== "string") throw result;
+    return result;
+  }
+
+  async function signTransactionWC(transactionRequestWC: any) {
+    // parse params from transactionRequestWC
+    const requestParams = parseExtendedJson(JSON.stringify(transactionRequestWC.value.params.request.params));
+    const txDetails:TransactionCommon = requestParams.transaction;
+    const sourceOutputs = requestParams.sourceOutputs as (Input & Output & ContractInfo)[]
+
+    const privateKey = store?.wallet?.privateKey;
+    const pubkeyCompressed = store?.wallet?.publicKeyCompressed
+    if(!privateKey || !pubkeyCompressed) return
+
+    // prepare libauth template for input signing
+    const template = importWalletTemplate(walletTemplateP2pkhNonHd);
+    if (typeof template === "string") throw new Error("Transaction template error");
+
+    // configure compiler
+    const compiler = walletTemplateToCompilerBCH(template);
+
+    const txTemplate = {...txDetails} as TransactionTemplateFixed<typeof compiler>;
+
+    for (const [index, input] of txTemplate.inputs.entries()) {
+      const sourceOutputsUnpacked = sourceOutputs;
+      if (sourceOutputsUnpacked[index].contract?.artifact.contractName) {
+        // instruct compiler to produce signatures for relevant contract inputs
+
+        // replace pubkey and sig placeholders
+        let unlockingBytecodeHex = binToHex(sourceOutputsUnpacked[index].unlockingBytecode);
+        const sigPlaceholder = "41" + binToHex(Uint8Array.from(Array(65)));
+        const pubkeyPlaceholder = "21" + binToHex(Uint8Array.from(Array(33)));
+        if (unlockingBytecodeHex.indexOf(sigPlaceholder) !== -1) {
+          // compute the signature argument
+          const hashType = SigningSerializationFlag.allOutputs | SigningSerializationFlag.utxos | SigningSerializationFlag.forkId;
+          const context: CompilationContextBCH = { inputIndex: index, sourceOutputs: sourceOutputsUnpacked, transaction: txDetails };
+          const signingSerializationType = new Uint8Array([hashType]);
+
+          const coveredBytecode = sourceOutputsUnpacked[index].contract?.redeemScript;
+          if (!coveredBytecode) {
+            alert("Not enough information provided, please include contract redeemScript");
+            return;
+          }
+          const sighashPreimage = generateSigningSerializationBCH(context, { coveredBytecode, signingSerializationType });
+          const sighash = hash256(sighashPreimage);
+          const signature = secp256k1.signMessageHashSchnorr(privateKey, sighash);
+          if (typeof signature === "string") {
+            alert(signature);
+            return;
+          }
+          const sig = Uint8Array.from([...signature, hashType]);
+
+          unlockingBytecodeHex = unlockingBytecodeHex.replace(sigPlaceholder, "41" + binToHex(sig));
+        }
+        if (unlockingBytecodeHex.indexOf(pubkeyPlaceholder) !== -1) {
+          unlockingBytecodeHex = unlockingBytecodeHex.replace(pubkeyPlaceholder, "21" + binToHex(pubkeyCompressed));
+        }
+
+        input.unlockingBytecode = hexToBin(unlockingBytecodeHex);
+      } else {
+        // replace unlocking bytecode for non-contract inputs having placeholder unlocking bytecode
+        const sourceOutput = sourceOutputsUnpacked[index];
+        if (!sourceOutput.unlockingBytecode?.length && toCashaddr(sourceOutput.lockingBytecode) === store.wallet?.getDepositAddress()) {
+          input.unlockingBytecode = {
+            compiler,
+            data: {
+              keys: { privateKeys: { key: privateKey } },
+            },
+            valueSatoshis: sourceOutput.valueSatoshis,
+            script: "unlock",
+            token: sourceOutput.token,
+          }
+        }
+      }
+    };
+
+    // generate and encode transaction
+    const generated = generateTransaction(txTemplate);
+    if (!generated.success) throw Error(JSON.stringify(generated.errors, null, 2));
+
+    const encoded = encodeTransaction(generated.transaction);
+    const hash = binToHex(sha256.hash(sha256.hash(encoded)).reverse());
+    const signedTxObject = { signedTransaction: binToHex(encoded), signedTransactionHash: hash };
+
+    // send transaction
+    const { id, topic } = transactionRequestWC.value;
+    const response = { id, jsonrpc: '2.0', result: signedTxObject };
+    if (requestParams.broadcast) {
+      try{
+        Notify.create({
+          spinner: true,
+          message: 'Sending transaction...',
+          color: 'grey-5',
+          timeout: 750
+        })
+        const txId = await store.wallet?.submitTransaction(hexToBin(signedTxObject.signedTransaction));
+        const alertMessage = `Sent WalletConnect transaction '${requestParams.userPrompt}'`
+        Dialog.create({
+          component: alertDialog,
+          componentProps: {
+            alertInfo: { message: alertMessage, txid: txId as string } 
+          }
+        })
+      } catch(error){
+        console.log(error)
+        Notify.create({
+          type: 'negative',
+          message: typeof error == 'string' ? error :  "Error in sending transaction"
+        })
+        return
+      }   
+    }
+    await web3wallet.value?.respondSessionRequest({ topic, response });
+    // close dialog when broadcast = false
+    // showDialog.value = false
+    // TODO: rework
+    // emit('signedTransaction', requestParams.broadcast);
+  }
+
+  async function rejectTransaction(transactionRequestWC: any){
+    const { id, topic } = transactionRequestWC.value;
+    const response = { id, jsonrpc: '2.0', error: getSdkError('USER_REJECTED') };
+    await web3wallet.value?.respondSessionRequest({ topic, response });
+    // TODO: rework
+    // emit('rejectTransaction')
+  }
 
   return { web3wallet, activeSessions, initweb3wallet, wcRequest }
 })
